+++ src/store/settingsStore.js
@@ -42,6 +42,8 @@ export const useSettingsStore = defineStore('settings', {
       gain: 0,
       offset: 0,
       useSolve: false,
+      imageScale: 100,
+      imageQuality: 90,
     },
     flats: {
       selectedOption: 'AutoExposure',